# services/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AddToWatchlistRequest:
    properties:
      athlete_id:
        type: string
      notes:
        type: string
      priority_tag_id:
        type: string
    required:
    - athlete_id
    type: object
...
  dto.AthleteProfileResponse:
    properties:
      age:
        type: integer
      athlete_id:
        type: string
      created_at:
        type: string
      current_academy_id:
        type: string
      current_academy_name:
        type: string
      highest_competition_level:
        type: string
      id:
        type: string
      leaderboard_rank:
        type: integer
      leaderboard_score:
        type: number
      sport_id:
        type: string
      total_academies:
        type: integer
      total_competitions:
        type: integer
      updated_at:
        type: string
      years_active:
        type: integer
    type: object
...
  dto.CreateScoutProfileRequest:
    properties:
      avatar_attachment_id:
        type: string
      bio:
        type: string
      organization_name:
        type: string
    required:
    - organization_name
    type: object
...
  dto.LogScoutActivityRequest:
    properties:
      action:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
    required:
    - action
    type: object
...
  dto.ScoutProfileResponse:
    properties:
      avatar_attachment_id:
        type: string
      bio:
        type: string
      created_at:
        type: string
      id:
        type: string
      organization_name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdateScoutProfileRequest:
    properties:
      avatar_attachment_id:
        type: string
      bio:
        type: string
      organization_name:
        type: string
    type: object
...
  dto.UpdateWatchlistEntryRequest:
    properties:
      notes:
        type: string
      priority_tag_id:
        type: string
    type: object
...
  dto.WatchlistEntryResponse:
    properties:
      added_at:
        type: string
      athlete_id:
        type: string
      athlete_profile:
        $ref: '#/definitions/dto.AthleteProfileResponse'
      id:
        type: string
      notes:
        type: string
      priority_tag_id:
        type: string
      scout_id:
        type: string
      updated_at:
        type: string
    type: object
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: dev@sportsentral.id
    name: Sport-Sentral Development Team
  description: Centralized API Documentation for Sport-Sentral Microservices Gateway
    (Auth, RBAC, Academy, Venue, Sport, Competition, Scout, Attachment, Log).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sport-Sentral API Gateway
  version: "1.0"
paths:
...
  /scouts/athletes:
    get:
      description: Search athlete profile directory filtered by sport, age range,
        and highest competition level.
      parameters:
      - description: Filter Sport ID (UUID)
        in: query
        name: sport_id
        type: string
      - description: Minimum age
        in: query
        name: min_age
        type: integer
      - description: Maximum age
        in: query
        name: max_age
        type: integer
      - description: Highest competition level
        in: query
        name: level
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Search and filter athlete profiles
      tags:
      - Scouts
  /scouts/athletes/{id}:
    get:
      description: Retrieve career statistics, leaderboard rankings, competition history,
        and academy background for an athlete.
      parameters:
      - description: Athlete User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Sport ID (UUID)
        in: query
        name: sport_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AthleteProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get athlete performance and statistics profile
      tags:
      - Scouts
  /scouts/leaderboard:
    get:
      description: Retrieve top athlete rankings calculated from official tournament
        score weighting per sport.
      parameters:
      - description: Sport ID (UUID)
        in: query
        name: sport_id
        required: true
        type: string
      - description: Period Tag ID (UUID)
        in: query
        name: period_tag_id
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get athlete leaderboard rankings
      tags:
      - Scouts
  /scouts/logs:
    get:
      description: Retrieve activity logs performed by a talent scout.
      parameters:
      - description: Scout ID (UUID)
        in: query
        name: scout_id
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get talent scout activity history
      tags:
      - Scouts
    post:
      consumes:
      - application/json
      description: Record scout activity history when viewing/analyzing athlete profiles.
      parameters:
      - description: Scout ID (UUID)
        in: query
        name: scout_id
        required: true
        type: string
      - description: Activity Log Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogScoutActivityRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Log talent scout activity
      tags:
      - Scouts
  /scouts/profile:
    get:
      description: Retrieve scout profile data for the logged-in account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ScoutProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get own talent scout profile
      tags:
      - Scouts
    post:
      consumes:
      - application/json
      description: Create a new scout profile linked to the logged-in user account.
      parameters:
      - description: Scout Profile Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateScoutProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ScoutProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create talent scout profile
      tags:
      - Scouts
  /scouts/profile/{id}:
    put:
      consumes:
      - application/json
      description: Update scout organization, bio, or avatar information.
      parameters:
      - description: Scout ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Scout Profile Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateScoutProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ScoutProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update talent scout profile
      tags:
      - Scouts
  /scouts/watchlist:
    get:
      description: Retrieve list of athletes monitored by scout with pagination.
      parameters:
      - description: Scout ID (UUID)
        in: query
        name: scout_id
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get athlete watchlist
      tags:
      - Scouts
    post:
      consumes:
      - application/json
      description: Add an athlete to talent scout watchlist with notes and priority.
      parameters:
      - description: Scout ID (UUID)
        in: query
        name: scout_id
        required: true
        type: string
      - description: Athlete Watchlist Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddToWatchlistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.WatchlistEntryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add athlete to watchlist
      tags:
      - Scouts
  /scouts/watchlist/{athlete_id}:
    delete:
      description: Remove an athlete from the scout watchlist.
      parameters:
      - description: Athlete ID (UUID)
        in: path
        name: athlete_id
        required: true
        type: string
      - description: Scout ID (UUID)
        in: query
        name: scout_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Remove athlete from watchlist
      tags:
      - Scouts
  /scouts/watchlist/{id}:
    put:
      consumes:
      - application/json
      description: Update notes or priority for an athlete in the watchlist.
      parameters:
      - description: Watchlist Entry ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Scout ID (UUID)
        in: query
        name: scout_id
        required: true
        type: string
      - description: Watchlist Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWatchlistEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.WatchlistEntryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update watchlist entry notes
      tags:
      - Scouts
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in the format: `Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"